manage dependencies

+++ uninstall.sh
@@ -7,24 +7,40 @@ SHARE_FILE1=".default_cursor"
 SHARE_FILE2=".default_session"
 SERVICE_FILE="ezs.service"
 
-# Eliminar el archivo binario
-[ -f "/usr/bin/$BIN_FILE" ] && sudo rm "/usr/bin/$BIN_FILE"
+# Remove the binary file
+if [ -f "/usr/bin/$BIN_FILE" ]; then
+    echo "Removing binary file: $BIN_FILE from /usr/bin/"
+    sudo rm "/usr/bin/$BIN_FILE"
+fi
 
-# Eliminar los archivos de configuración
-[ -f "$SHARE_DIR/$SHARE_FILE1" ] && sudo rm "$SHARE_DIR/$SHARE_FILE1"
-[ -f "$SHARE_DIR/$SHARE_FILE2" ] && sudo rm "$SHARE_DIR/$SHARE_FILE2"
+# Remove the configuration files
+if [ -f "$SHARE_DIR/$SHARE_FILE1" ]; then
+    echo "Removing session data file: $SHARE_FILE1 from $SHARE_DIR"
+    sudo rm "$SHARE_DIR/$SHARE_FILE1"
+fi
 
-# Eliminar el directorio de configuración si está vacío
-[ -d "$SHARE_DIR" ] && [ -z "$(ls -A $SHARE_DIR)" ] && sudo rmdir "$SHARE_DIR"
+if [ -f "$SHARE_DIR/$SHARE_FILE2" ]; then
+    echo "Removing session data file: $SHARE_FILE2 from $SHARE_DIR"
+    sudo rm "$SHARE_DIR/$SHARE_FILE2"
+fi
 
-# Eliminar el archivo de servicio
-[ -f "/etc/systemd/system/$SERVICE_FILE" ] && {
+# Remove the configuration directory if empty
+if [ -d "$SHARE_DIR" ] && [ -z "$(ls -A $SHARE_DIR)" ]; then
+    echo "Removing directory: $SHARE_DIR"
+    sudo rmdir "$SHARE_DIR"
+fi
+
+# Remove the service file
+if [ -f "/etc/systemd/system/$SERVICE_FILE" ]; then
+    echo "Stopping and disabling service: $SERVICE_FILE"
     sudo systemctl stop "$SERVICE_FILE"
     sudo systemctl disable "$SERVICE_FILE"
+    echo "Removing service file: $SERVICE_FILE from /etc/systemd/system/"
     sudo rm "/etc/systemd/system/$SERVICE_FILE"
-}
+fi
 
-# Recargar el daemon de systemd
+# Reload the systemd daemon
+echo "Reloading systemd daemon"
 sudo systemctl daemon-reload
 
-echo "completed"
+echo "Uninstallation completed."
